Readme information
We are task to create an application that receives input from the user for a recipe for Lindiwe’s birthday party. The application prints out information based on the user pressing a string value to continue with the program. If the user presses a button that is not is string the error message. Most of the input command are “Press enter” to continue with the program, you can also press spacebar to continue. The ingredient is in array and they will be display as you continue with the project. When you get to the section where you are supposing to choose 1 option. You can only enter int data the rest will give you a error so it you are going through the program read what the command is asking you to input. After that the program will run as usual. You will run into a reset ingredient option which will direct you to an option where you will be required to type in the new ingredients. Those ingredient will be stored in the method then you will be able to display those new ingredients.